refactor: split api client by domain
